+++ src/plonemeeting/portal/core/viewlets/overrides.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
-
+from plone.app.layout.analytics.view import AnalyticsViewlet
 from plone.app.layout.viewlets.content import DocumentBylineViewlet
+from plonemeeting.portal.core.content.institution import IInstitution
 from plonemeeting.portal.core.utils import is_decisions_manager
 from plonemeeting.portal.core.utils import is_publications_manager
 from Products.CMFCore.permissions import ManagePortal
@@ -20,3 +21,16 @@ class PMDocumentBylineViewlet(DocumentBylineViewlet):
     def show_modification_date(self):
         """Show modified no matter it was published, useful for Publication."""
         return True
+
+class PMAnalyticsViewlet(AnalyticsViewlet):
+
+    @property
+    def webstats_js(self):
+        webstats = super().webstats_js or ""
+        utils_view = self.context.restrictedTraverse("@@utils_view")
+        if utils_view.is_in_institution():
+            # If the context is inside an institution, we want to add the institution's webstats JS
+            institution = utils_view.get_current_institution()
+            if hasattr(institution, 'webstats_js') and institution.webstats_js:
+                webstats += institution.webstats_js
+        return webstats